Refactored the landOnField method for ownable classes into property
